+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Ignore case of the `merge-mode` attribute [#1023](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1023)
 - Improved xsd mapping for `ea:attribute` [#1024](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1024)
 - Added code completion for `list-view:column`:`class` [#1028](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1028)
+- Added code completion for `custom-section` & `custom-attribute`:`class` [#1029](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1029)
 
 ## [2024.1.0]